# sample-api-documentation.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Megaexe Post Management",
      "description": "Megaexe project task",
      "version": "1.0",
      "contact": {
  
      }
    },
...
    "servers": [],
    "components": {
      "securitySchemes": {
        "bearer": {
          "scheme": "bearer",
          "bearerFormat": "JWT",
          "type": "http"
        }
      },
      "schemas": {
...
        "CreateUserDto": {
          "type": "object",
          "properties": {
            "role": {
              "type": "string"
            },
            "first_name": {
              "type": "string"
            },
            "last_name": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            },
            "image": {
              "type": "string"
            },
            "phone_number": {
              "type": "string"
            },
            "gender": {
              "type": "string"
            }
          },
          "required": [
            "role",
            "first_name",
            "last_name",
            "email",
            "image",
            "phone_number"
          ]
        },
        "UpdateUserDto": {
          "type": "object",
          "properties": {
            "role_id": {
              "type": "string"
            },
            "permission_id": {
              "type": "string"
            },
            "first_name": {
              "type": "string"
            },
            "last_name": {
              "type": "string"
            },
            "username": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "image": {
              "type": "string"
            },
            "phone_number": {
              "type": "string"
            },
            "home_address": {
              "type": "string"
            },
            "state_of_residence": {
              "type": "string"
            },
            "lga": {
              "type": "string"
            },
            "geo_political_zone": {
              "type": "string"
            }
          },
          "required": [
            "role_id",
            "permission_id",
            "first_name",
            "last_name",
            "username",
            "email",
            "image",
            "phone_number"
          ]
        },
...
        "CreateVotingDto": {
          "type": "object",
          "properties": {
            "post_id": {
              "type": "string"
            },
            "vote_type": {
              "type": "string"
            }
          },
          "required": [
            "post_id",
            "vote_type"
          ]
        },
...
      }
    },
    "paths": {
...
      "/api/v1/users": {
        "post": {
          "operationId": "create",
          "parameters": [],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserDto"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": ""
            }
          },
          "tags": [
            "User Managements"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        },
        "get": {
          "operationId": "findAll",
          "parameters": [
            {
              "name": "page",
              "required": true,
              "in": "query",
              "schema": {
                "type": "number"
              }
            },
            {
              "name": "per_page",
              "required": true,
              "in": "query",
              "schema": {
                "type": "number"
              }
            },
            {
              "name": "query",
              "required": true,
              "in": "query",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "from",
              "required": true,
              "in": "query",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "to",
              "required": true,
              "in": "query",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "User Managements"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        }
      },
      "/api/v1/users/search": {
        "get": {
          "operationId": "search",
          "parameters": [
            {
              "name": "query",
              "required": true,
              "in": "query",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "per_page",
              "required": true,
              "in": "query",
              "schema": {
                "type": "number"
              }
            }
          ],
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "User Managements"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        }
      },
      "/api/v1/users/{id}": {
        "get": {
          "operationId": "findOne",
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "User Managements"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        },
        "patch": {
          "operationId": "update",
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "User Managements"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        }
      },
...
      "/api/v1/voting": {
        "post": {
          "operationId": "create",
          "parameters": [],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateVotingDto"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": ""
            }
          },
          "tags": [
            "Voting Management"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        },
        "get": {
          "operationId": "findAll",
          "parameters": [
            {
              "name": "page",
              "required": true,
              "in": "query",
              "schema": {
                "type": "number"
              }
            },
            {
              "name": "per_page",
              "required": true,
              "in": "query",
              "schema": {
                "type": "number"
              }
            },
            {
              "name": "user_id",
              "required": true,
              "in": "query",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "vote_type",
              "required": true,
              "in": "query",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "Voting Management"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        }
      },
      "/api/v1/voting/{id}": {
        "get": {
          "operationId": "findOne",
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "Voting Management"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        }
      },
...
    }
  }